(Fix) 삭제 기능 추가 및 서비스 수정 🐛

+++ src/main/java/com/seven/marketclip/chat/service/ChatRoomService.java
@@ -92,6 +92,9 @@ public class ChatRoomService {
     @Transactional  //채팅방 check box 삭제 API 4번
     public void removeChatRoom(List<String> listChatRoomId){
         for (String chatRoomId:listChatRoomId) {
+            if(chatRoomRepository.findById(chatRoomId).isEmpty()){
+                throw new CustomException(CHAT_ROOM_NOT_FOUND);
+            }
             chatRoomRepository.deleteById(chatRoomId);
         }
     }
